+++ CHANGELOG.md
@@ -6,9 +6,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.0.3] - 2019-10-12
+### Added
+- Solution to Problem 3 - **Largest Prime Factor**
+- utility function: `GetFactors( int ) []int`
+- utility function: `IsPrime( int ) bool`
+
 ## [0.0.2] - 2019-10-11
 ### Added
 - Solution to Problem 2 - **Even Fibonacci numbers**
+
 
 ## [0.0.1] - 2019-10-10
 ### Added